refactor: change structure
delete unused component, rename vars
