+++ src/sender/game/index.ts
@@ -3,4 +3,4 @@ export * from './Attack';
 export * from './RandomAttack';
 export * from './StartGame';
 export * from './Killed';
-export * from './WhoMove';
+export * from './turnUser';
